Made ID to a unique account ID and made sure it is in the constructors.

+++ RebelAllianceBank/Interfaces/IBankAccount.cs
@@ -2,9 +2,8 @@
 {
     public interface IBankAccount
     {
-        public int ID { get; set; }
+        public long ID { get; set; }
         public string UserId { get; set; }
-        //public long AccountNr { get; set; }
         public int AccountType { get; set; }
         public string AccountName { get; set; }
         public decimal Balance { get; set; }
